added add_pseudocount to qnorm_bs_files

+++ src_for_distrib/drivers/do_peak_and_epod_calls.py
@@ -28,6 +28,33 @@ import hdf_utils
 import anno_tools as anno
 import peak_utils as pu
 
+def choose_threshold(idr_files, cutoffs, best_thresh_path, mean_fname, samp, score_type):
+
+    if score_type == 'log10p':
+        lowest = 0
+        highest = 20
+    elif score_type == 'rz':
+        lowest = -10
+        highest = 11
+
+    # I need to add the logic to determine whether samples were paired. If not, just
+    #  use the only replicate's data, and no idr_files.
+    cmd = SCORE_SCRIPT.format(
+        ' '.join(idr_files), # space-separated file names for input files
+        ' '.join([str(c) for c in cutoffs]), # space-separated thresholds
+        best_thresh_path,
+        lowest,
+        highest,
+        mean_fname,
+    )
+    retcode = subprocess.call(cmd, shell=True)
+    if retcode != 0:
+        print("\n*******************")
+        print("Encountered an error in determining best peak calling threshold for sample {}.".format(samp))
+        print("Attempted command was:\n{}".format(cmd))
+        print("*******************\n")
+
+
 
 def call_peaks(in_fname, out_path, cutoff, samp_name):
     '''Utility function used to wrap creation of peak calling subprocess
@@ -247,149 +274,7 @@ def calc_idr(paired, out_files, ctg_lut, out_path,
 
         return final_fname,placeholder_files
 
-def get_kl_divergences(peak_scores, nonpeak_scores,
-                    lowest_bin=-10, highest_bin=11, alpha=0.05, n_b=1000):
-    """Calculate KL divergence between peaks and non-peaks. Also
-    performs bootstrapping to get the upper and lower confidence limits
-    for the KL divergence.
-    """
 
-    if peak_scores.size == 0:
-        return (0,0,0)
-    elif peak_scores.size == 1:
-        peak_scores = np.expand_dims(peak_scores, -1)
-
-    bins = np.arange(lowest_bin, highest_bin)
-    counts_1,bins_1 = np.histogram(peak_scores, bins=bins)
-    counts_2,bins_2 = np.histogram(nonpeak_scores, bins=bins)
-
-    entropy=scipy.stats.entropy( counts_1+1, counts_2+1 )
-
-    value_distr_1 = iid_bootstrap(peak_scores, replications=n_b, replace=True)
-    value_distr_2 = iid_bootstrap(nonpeak_scores, replications=n_b, replace=True)
-
-    entropy_from_bootstrap=[]
-
-    for i in range(n_b):
-        counts_1, bins_1 = np.histogram(
-            value_distr_1[i,:],
-            bins=bins,
-        )
-        counts_2, bins_2 = np.histogram(
-            value_distr_2[i,:],
-            bins=bins,
-        )
-        entropy_boot = scipy.stats.entropy( counts_1+1, counts_2+1 )
-        entropy_from_bootstrap.append( entropy_boot )
-
-    # now get the confidence interval based on percentiles
-    cl_lo = scipy.stats.scoreatpercentile( entropy_from_bootstrap, 100*(alpha/2) )
-    cl_hi = scipy.stats.scoreatpercentile( entropy_from_bootstrap, 100*(1- (alpha/2)) )
-
-    return (cl_lo, entropy, cl_hi)
-
-           
-def choose_final_threshold(idr_files, ctg_lut, spike_name, mean_fname,
-                        lowest_bin=-10, highest_bin=11, alpha=0.05, n_b=1000):
-    """Get the cutoff index that we consider to be the best cutoff for
-    robustly distinguishing between peak and non-peak regions of the genome.
-    """
-
-    divergences = []
-    print(idr_files)
-    for i,idr_fname in enumerate(idr_files):
-        print("=============================")
-        print("Calculating KL divergence between peaks in {} and non-peaks.".format(idr_fname))
-        print("=============================")
-
-        final_peaks = anno.NarrowPeakData()
-        final_peaks.parse_narrowpeak_file(idr_fname)
-        complement_bed_data = final_peaks.fetch_complement_bed_data(
-            contig_lut = ctg_lut,
-            filter_chrs = [spike_name],
-        )
-
-        tmp_dir = tempfile.TemporaryDirectory()
-
-        # write the bed file containing peak-less regions to bed file
-        complement_bed_data.fname = '/home/schroedj/nopeak_data.bed'
-        complement_bed_data.write_file()
-
-        nonpeak_bed_fname = os.path.join(tmp_dir.name, 'nopeak_data.bed')
-        complement_bed_data.fname = nonpeak_bed_fname
-        complement_bed_data.write_file()
-
-        # write mean peak and non-peak scores to temporary files
-        peak_score_fname = os.path.join(tmp_dir.name, 'peakscore.bed')
-        #peak_score_fname = '/home/schroedj/peakscore.bed'
-        nonpeak_score_fname = os.path.join(tmp_dir.name, 'nonpeakscore.bed')
-        #nonpeak_score_fname = '/home/schroedj/nonpeakscore.bed'
-
-        #if not os.path.isfile(final_peaks.fname):
-        #    raise()
-        #if not os.path.isfile(mean_fname):
-        #    raise()
-        #if not os.path.isfile(nopeak_bed_fname):
-        #    raise()
-        #if not os.path.isdir(tmp_dir.name):
-        #    raise()
-        bed_map_cmd = "bedtools map \
-                -a {} \
-                -b {} \
-                -o mean -c 4 \
-                | cut -f 7 \
-                > {}".format(
-                final_peaks.fname,
-                mean_fname,
-                peak_score_fname,
-        )
-        print(bed_map_cmd)
-        subprocess.call(
-            bed_map_cmd,
-            shell = True,
-        )
-        peak_scores = np.loadtxt(peak_score_fname)
-        
-        bed_map_cmd = "bedtools map \
-                -a {} \
-                -b {} \
-                -o mean -c 4 \
-                | cut -f 7 \
-                > {}".format(
-                nonpeak_bed_fname,
-                mean_fname,
-                nonpeak_score_fname,
-        )
-        print(bed_map_cmd)
-        subprocess.call(
-            bed_map_cmd,
-            shell = True,
-        )
-        nonpeak_scores = np.loadtxt(nonpeak_score_fname)
-
-        tmp_dir.cleanup()
-        print(nonpeak_scores)
-
-        # give a tuple of (lower_cl, observed, upper_cl) for the KL divergence
-        this_div_info = get_kl_divergences(
-            peak_scores,
-            nonpeak_scores,
-            lowest_bin,
-            highest_bin,
-            alpha,
-            n_b,
-        )
-        divergences.append(this_div_info)
-    # make divergences an array, and transpose it and slice rows in reverse
-    #  so final array's rows are upper, observed, lower at indices 0,1,2, repectively,
-    #  and columns are cutoffs
-    div_arr = np.asarray(divergences).T[::-1,:]
-    max_observed = div_arr[1,:].max()
-    # get max index at which upper cl is greater than the max observed KL divergence
-    cutoff_idx = np.where(div_arr[0,:] > max_observed)[0].max()
-    return (cutoff_idx, div_arr, max_observed)
-    
- 
 def process_sample(line, conf_dict_global, invert):
 
     dirname,samp_conf = line.rstrip().split()
@@ -466,6 +351,10 @@ def process_sample(line, conf_dict_global, invert):
 
                 # we'll collect the IDR-passing peaks' narrowpeak files in this list
                 idr_files = []
+                # one_rep is used later to determine whether to use idr-passing peaks, or just a single replicate's peaks in threshold choosing.
+                one_rep = False
+                # if there's one repicate, we use this list later for threshold choosing
+                one_rep_files = []
                 # loop over multiple score cutoffs.
                 for cutoff in these_cutoffs:
 
@@ -484,12 +373,13 @@ def process_sample(line, conf_dict_global, invert):
 
                     if paired:
                         if len(fname_list) < 2:
+                            one_rep = True
+                            one_rep_files.extend(out_files)
                             print("============================")
                             print("Only one replicate found in sample {}, skipping IDR calculation for peaks.".format(samp))
                             print("----------------------------")
                             continue
 
-    
                     idr_fname,placeholder_files = calc_idr(
                         paired,
                         out_files,
@@ -510,60 +400,34 @@ def process_sample(line, conf_dict_global, invert):
                         pathlib.Path(idr_fname).touch()
                         placeholder_files.append(idr_fname)
 
-                if score_type == 'log10p':
-                    lowest = 0
-                    highest = 20
-                elif score_type == 'rz':
-                    lowest = -10
-                    highest = 11
-
-                best_cutoff_idx,kl_divs,max_obs_kl_div = choose_final_threshold(
-                    idr_files,
-                    ctg_lut,
-                    spike_name,
-                    mean_fname,
-                    lowest,
-                    highest,
-                )
-                best_cutoff = these_cutoffs[best_cutoff_idx]
-                best_result = idr_files[best_cutoff_idx]
-                basename = os.path.basename(best_result)
-                shutil.copy(best_result, os.path.join(best_thresh_path, basename))
-
-                plt_name = os.path.join(
-                    best_thresh_path,
-                    "kl_div_cutoff_{}.png".format(basename),
-                )
-
-                err_bar_arr = kl_divs[[2,0],:] - kl_divs[1,:][None,:]
-                err_bar_arr[0,:] *= -1
-
-                fig, ax = plt.subplots(figsize=(9,5))
-                ax.axhline(
-                    y=kl_divs[1,:].max(),
-                    color='r',
-                    linestyle='--',
-                )
-                ax.errorbar(
-                    x = these_cutoffs,
-                    y = kl_divs[1,:],
-                    yerr = err_bar_arr,
-                    color='tab:blue',
-                    ecolor='tab:blue',
-                )
-                ax.set_xlabel('threshold')
-                ax.set_ylabel('KL divergence')
-                ax.spines['top'].set_visible(False)
-                ax.spines['right'].set_visible(False)
-
-                plt.savefig(plt_name)
-                plt.close()
+                # if there was more than one threshold, choose the best one
+                if len(these_cutoffs) > 1:
+                    if one_rep:
+                        choose_threshold(
+                            one_rep_files,
+                            these_cutoffs,
+                            best_thresh_path,
+                            mean_fname,
+                            samp,
+                            score_type,
+                        )
+                    else:
+                        choose_threshold(
+                            idr_files,
+                            these_cutoffs,
+                            best_thresh_path,
+                            mean_fname,
+                            samp,
+                            score_type,
+                        )
 
                 # if we made any empty narrowpeak files, delete them now.
-                if len(placeholder_files) > 0:
-                    for fname in placeholder_files:
-                        print("Removing empty placeholder file {}.".format(fname))
-                        os.remove(fname)
+                if not one_rep:
+                    if len(these_cutoffs) > 1:
+                        if len(placeholder_files) > 0:
+                            for fname in placeholder_files:
+                                print("Removing empty placeholder file {}.".format(fname))
+                                os.remove(fname)
 
             # do epod calling
             if not 'epods' in skipsteps:
@@ -582,6 +446,9 @@ def process_sample(line, conf_dict_global, invert):
                     # get replicate fnames
                     fname_search = fname.format("rep*")
                     rep_fname_list = glob.glob(os.path.join(in_path, fname_search))
+                    if len(rep_fname_list) == 0:
+                        print(fname_search)
+                        print(in_path)
 
                     strict_epod_outfiles = []
                     loose_epod_outfiles = []
@@ -680,6 +547,7 @@ if __name__ == "__main__":
     conf_file = args.main_conf
     conf_dict_global = toml.load(conf_file)
 
+    SPIKE = conf_dict_global["genome"]["spike_in_name"]
     BASEDIR = conf_dict_global["general"]["basedir"]
     BINDIR = conf_dict_global["general"]["bindir"]
     RESOLUTION = conf_dict_global["genome"]["resolution"]
@@ -707,6 +575,16 @@ if __name__ == "__main__":
                            --plot --log-output-file {}.log --verbose\
                            --output-file {}"
     
+    SCORE_SCRIPT = "python {}/peakcalling/choose_threshold.py\
+                        --infiles {{}}\
+                        --thresholds {{}}\
+                        --out_dir {{}}\
+                        --lowest {{}}\
+                        --highest {{}}\
+                        --mean_score_fname {{}}\
+                        --ref_db {}\
+                        --spikein_name {}".format(BINDIR, SEQ_DB, SPIKE)
+
     ## this one just need the peaks .gr file and output prefix
     #OVERLAP_SCRIPT = "python {}/peakcalling/analyze_peaks.py {{}}\
     #                  /data/petefred/st_lab_work/e_coli_data/regulondb_20180516/BindingSites_knownsites_flags.gr > \
